move test_protocol_error
